403页面

+++ src/main/java/com/dj/bank/pojo/BankLoans.java
@@ -47,7 +47,7 @@ public class BankLoans {
     /**
      * 剩余还款月
      */
-    private BigDecimal payMonthNumber;
+    private Integer payMonthNumber;
 
     /**
      * 1在，2无
